+++ CHANGELOG.md
@@ -137,6 +137,12 @@
   [#3745](https://github.com/Kong/kong-operator/pull/3745)
 - Added support for using `KonnectGatewayControlPlane` of type group with `KonnectExtension`s.
   [#3711](https://github.com/Kong/kong-operator/pull/3711)
+- `MCPServer`: set `KongRoute` and `KongService` Konnect IDs during
+  reconciliation.
+  [#3904](https://github.com/Kong/kong-operator/pull/3904)
+- `MCPServer`: set MCPServer ID as the name identifier for the Kubernetes resource and store
+  the Konnect name in the `MCPServer` CRD status.
+  [#3904](https://github.com/Kong/kong-operator/pull/3904)
 
 ### Changed